smart-plant-watering/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
scikit-learn
twilio
python-dotenv
